requirements.txt
python-telegram-bot==21.9
playwright==1.49.1
playwright-stealth==1.0.6
anthropic==0.42.0
pydantic==2.10.4
pyyaml==6.0.2
loguru==0.7.3
yoyo-migrations==9.0.0
aiohttp==3.11.11
pytest==8.3.4
pytest-asyncio==0.26.0
